requirements.txt
pydantic>=2.0
pyyaml>=6.0
rich>=13.0
tqdm>=4.0
numpy>=1.24
pandas>=2.0
matplotlib>=3.8
